User.java class is fully complete
