--- app/src/main/java/nl/tudelft/b_b_w/Objects/Public_Key.java
@@ -1,15 +1,0 @@
-package nl.tudelft.b_b_w.Objects;
-
-/**
- * Class for creating aan acquaintance,
- * this is someone who you paired with but is not in your contact list yet.
- * This is a data transfer object.
- */
-
-public class Public_Key {
-
-
-
-
-
-}
