+++ app/src/main/java/nl/tudelft/b_b_w/blockchaincomponents/Public_Key.java
@@ -1,0 +1,10 @@
+package nl.tudelft.b_b_w.blockchaincomponents;
+
+/**
+ * Dummy class until the real public key class is merged
+ */
+
+public class Public_Key {
+
+
+}
