+++ app/src/test/java/nl/tudelft/b_b_w/blockchaincomponents/BlockTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.b_b_w.Objects;
+package nl.tudelft.b_b_w.blockchaincomponents;
 
 import org.junit.Before;
 import org.junit.Test;
